Reload wpml-config.xml to local variable
This changes the behavior of loading the wpml-config.xml to get refreshed before the post is saved.
Originally the config file gets loaded on init() and stays in this state until the class gets reinitialized. This leads to the problem of outdated translations options, when changing wpml-config.xml before save. To overcome this problem it is necessary to reload the wpml-config.xml in the beginning of the following methods:

- copy_post_metas
- copy_term_metas
- translate_types
- translate_taxonomies

The function to load the config file is put into the function load_wpml_config_file and added to the end of the file.